Assume cros_debug if nonchrome firmware
When running ChromiumOS on non-ChromeOS devices, the system will
never be in debug mode because the firmware fields that ChromeOS
expects will not be present.  Under these conditions, the system
will assume cros_debug is disabled, and crosh shell will be
unavailable.  This changes the behavior to assume that if you are
running ChromiumOS on non-ChromeOS hardware, you will want
cros_debug enabled.

BUG=None
TEST=Install ChromiumOS on a device with Intel firmware and ensure
     that the crosh "shell" command is available.
BRANCH=None

Change-Id: I37839649a31fb95f3791d2bd67f8ceca91902be4