Update patch set 1
Patch Set 1: Verified

Tested and verified that it builds, runs, has the indicated benefit.

This also is one of the items that helps pave the way to
evolve to EFI boot.  This would be nice since some
combinations of system-and-sata-drive don't enumerate
the sata-drive always at /dev/sda or /dev/sdc.  This is
a "feature"; drive-letter enumeration stability is not
guaranteed or even necessarily repeatable with same
hardware.  This messes-up legacy-boot which is erroneously
depending on drive-letter selection of boot partition.

EFI boot lets you choose boot partition by the partition's
partition-uuid in the GPT partition-table regardless of
where a sata drive falls in drive-letter-enumeration.  So
it would be nice to merge this eventually as part of EFI
boot enabling.  Other patches to enable EFI boot are also
needed.

Patch-set: 1
Reviewer: Gerrit User 1002140 <1002140@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Label: Verified=+1