Enables MTD mode via command line option.
Enables MTD mode with a command line option of
-nand=<write page bytes>,<write pages per erase block>,
<erase block offset to fts partition>,<erase block size of fts partition>

The problem is that we need to know exactly where the fts store partition is
and what the layout of the flash chip is in order to be able to use it
properly. This will vary per-device however, so for flexibility and testing
purposes, this command line option allows specification of arbitrary layouts.

BUG=chromium:221745
BRANCH=cros/embedded
TEST=verified during development via prints

Change-Id: I88640e777d5ccd4ffd3b39e1f131dc1d362c6b2c