Update patch set 1
Patch Set 1:

Luigi - the e820 fields are the memory maps between the BIOS and kernel. When I first wrote this, our BIOS was EFI-based so those were all empty (and I don't remember seeing the fields that Kees is adding, either). The bootstub translated the EFI-format memory maps into the e820 format before jumping to the kernel.

I tried adding direct support for the EFI maps, but was met with a resounding "Oh, HELL no!" on the LKML.

Patch-set: 1