Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(2 inline comments)

A few nits.

flash_ts.c is going into the firmware, and appears to have lots of potential overflow/underflow conditions.  Before we hook it up to the vboot code we need to take care of those, or a MTD-formatted device could potentially be used as an attack on verified boot.

(+sumit to get someone from security to examine; I don't seem to be able to add will to the bug...)

Patch-set: 2
Label: Code-Review=-1