Fix recovery mode entry due to invalid FW hash.
If FW hash is invalid, it is possible for boot to continue as normal
because the shared recovery reason is not updated. Update the shared
recovery reason to ensure the bootloader knows to do a recovery boot.

This change affects only systems that do not support RO_NORMAL (ex.
Haswell systems).

BUG=chrome-os-partner:20518.
TEST=Run firmware_CorruptBothFwSigAB on Peppy, verify PASS.
BRANCH=None.
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>

Change-Id: I3b311416bf8c504deb4154e5850d967253c9dc17