Update patch set 1
Patch Set 1:

Your right, I did drop it in the wrong directory by mistake.

And as for the right place to call it.... I'm trying to figure that out as I go. This approach is based on the pattern around tag_image.sh and how it's called.

If sign_official_build.sh is better, I'll do that. It means no new instruction is needed, which simplifies everything.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0