Update patch set 3
Patch Set 3:

As long as the factory install shim is still bootable (which I think is true), I'm ok with it.

However, I suspect that because if you look at EFI config files,

 menuentry "Alternate USB Boot" {
   linux (hd0,3)/boot/vmlinuz   // ,3 = rootfs /boot
 menuentry "local image A" {
   linux $grubpartA/boot/vmlinuz quiet console=tty2

It seems to me that for non-EFI BIOS (i.e., syslinux) it uses kernels inside #12 (EFI) partition, but for UEFI BIOS (i.e., grub), it relies on /boot to start system.

I think you probably want to fix EFI boot menu (and postinst) before applying this commit, or remove the "factory" or "install" striping.

Patch-set: 3
Label: Commit-Queue=0