Proposed way of accessing GBB region
Currently U-Boot must load the GBB before calling into vboot. Since it takes
about 200ms to load the entire GBB, we save boot time by only loading what is
needed. This is done in U-Boot by using knowledge of how the internal vboot
code works.

For memory-constrained environments, such as early-firmware-selection on ARM,
we don't have enough memory to load the entire GBB. Additionally the current
gbb_data pointer approach expects the GBB to be loaded into a contiguous
memory region. We cannot honor that either.

This CL contains a proposed way of accessing the GBB from vboot, where each
access is passed back to U-Boot as a region of SPI flash to make available.
U-Boot allocates the memory, does the read and returns the data. Vboot can
then cache it for as long as U-Boot runs. It is not part of shared data,
since there is likely little benefit to passing GBB fragments through shared
data, and the additional complexity does not seem to be worth it.

BUG=chrome-os-partner:21115
BUG=chrome:205554
BRANCH=none
TEST=none, proposal only, will do full CL when agreed

Change-Id: Iec003d817933271b3e2aaaca00da7c64d5ef5fe6
Signed-off-by: Simon Glass <sjg@chromium.org>