Implementation of Region API
This has been generalized to the concept of a region, which could be a part
of the firmware image, or a disk, or something else.

Assuming this makes sense, it probably needs to be split into two CLs. Also
needs some tidy-up, comments, etc.

BUG=chrome-os-partner:21115
BRANCH=none
TEST=manual
FEATURES=test sudo -E  emerge vboot_reference

Change-Id: I2c19e9dc2ed602d0642bbf4f7d27f79fe9fad873
Signed-off-by: Simon Glass <sjg@chromium.org>