Update patch set 6
Patch Set 6:

each unittest can take like 5 minutes or so, which means cumulatively, it might be like 20 to 30.  you should see a dot on the screen each time it finishes.  also, when it first runs, it'll try to cache some large archives, so that might take a bit of time too.

you can tail the test.log file in the same dir to see if it's making progress.

Patch-set: 6
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0