Initialize USB prior to displaying developer mode screen.
If libpayload is built with USB keyboard support, it will need to
initialize the USB controller prior to accepting input. This can take a
while on some platforms, which leads to user input lag and/or missed
keystrokes since it will occur while the developer mode screen is up. To
prevent the bad user experience, pre-initialize USB before displaying
the developer screen.

BUG=chrome-os-partner:22006.
TEST=Manual. Boot Peppy in dev mode, verify that keyboard is responsive
to input immediately when dev screen is up and no keystrokes are lost.
BRANCH=FalcoPeppy

Change-Id: I4ba0b6fe55346dd8f656195fcfbe9acf07c49dcf