Update patch set 2
Patch Set 2:

Putting this in vboot does make it universal but that could conflict with whatever optimizations are being done in the bootloader.

It is unfortunate to take the hit for USB init every time given the common case of not using USB keyboard or booting from USB, but the current situation is very confusing when things don't work right.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0