Update patch set 8
Patch Set 8:

I rolled back to an implementation similar to #2.

We have unit tests that emulate disk detection / removal, counting calls to VbExDiskGetInfo to change the detection state. The previous version threw these tests off, because we make 2 initial calls to VbExDiskGetInfo (first call outside loop is ignored, except for delay period), even if disks are detected. This is harmless in practice, but I think it might be hard to justify changing the test for this case.

Note that we don't break any tests by making two calls to VbExDiskGetInfo before displaying the remove screen. If we did, I'd definitely change that test, because it is the exact behavior we intend to modify with this change.

Patch-set: 8