Update patch set 8
Patch Set 8: Code-Review+2

Hmm... this seems to be an implementation detail of the test, not something we'd really want to prevent (unless you think there's something bad about always polling twice). In general, I'd say you should rather change the test than let an overly narrow implementation that was definitely not intended to fail in this way drive your design decisions for the actual code. But it's also really not that important, so feel free to push this if you just want to get it done.

Patch-set: 8
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2