Update patch set 8
Patch Set 8:

I'll change the test if you think it's best.

Let me explain more about my reasoning for thinking to not change the test. Consider ver#6 of the change and how it behaves when media is inserted (ignore the case we're trying to fix, where no media is present, that case is fine). We will always do two polls before displaying the 'remove' screen, and before we just did one poll. It's totally fine to do two polls, and nothing bad can come of it, but I'd say it's still preferable to do one poll instead of two (even though the code is cleaner with two polls). Basically, I feel bad about modifying a test to allow me to make things function slightly worse. On the other hand, it will make the code slightly better, so it's a trade-off.

Patch-set: 8
Label: Code-Review=0