Update patch set 1
Patch Set 1:

I tried testing this by booting a U-Boot based image on Link, but ToT doesn't work so it's hard to tell if I broke anything. The U-Boot Makefile should pass it's own regparm=3 through so this should continue to be compatible, but I haven't concrete demonstrated that.

Since we're not using U-Boot on x86 any more and the machines that already do should be on branches, this seems like an acceptable risk to me.

Patch-set: 1