vboot: require an additional recovery button press
For devices that have a physical recovery button, we require
an additional recovery button press in the switch from
normal to developer mode.

This prevents an "evil keyboard" attack in which a USB keyboard
(or other USB device pretending to be a keyboard) sends a
control-D shortly after every boot.  In that situation, when
users power-on in recovery mode, they will be forced to dev
mode even if it was not their intention.  Other attacks become
easier at that point.

BUG=chrome-os-partner:21729
TEST=none yet
BRANCH=none

Change-Id: Ib986d00d4567c2d447f8bbff0e5ccfec94596aa7