Update patch set 1
Patch Set 1:

That's the what
"""TEST=Works, but DO NOT ARCHIVE."""
is for!

Working on the logic to be able to load an real image while in recovery mode.
Maybe we need a new mode:
- normal
- dev
- recovery
and
- recovery | dev

Patch-set: 1
Reviewer: Gerrit User 1104704 <1104704@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0