Update patch set 1
Patch Set 1:

Is there a ddoc for this? We don't have a lot of space in the TPM NVRAM, so I'm worried about pushing too much in there. Also, what mechanism should we use to expand this space in the future if we need to add more variables to it? How does the firmware choose between regular and TPM values?

Patch-set: 1