Update patch set 2
Patch Set 2:

Kees asked:
> Is there a ddoc for this? 

No, just the bug report. There's also issue 375770, which is tangentially related.


> We don't have a lot of space in the TPM NVRAM, so I'm worried about pushing too much in there. 

Agreed. I checked, though: on Link, this reduces the available NV space from 702 bytes to 640 bytes.


> Also, what mechanism should we use to expand this space in the future if we need to add more variables to it? 

This implementation saves a full copy of the VbNvContext raw bytes, just like the nvram does. But only certain fields are backed up - the others are zero. The current nvram storage has 47 unused bits, without even repurposing some of the field that would never need to be backed up. So we can add at least 47 bits without needing to worry.


> How does the firmware choose between regular and TPM values?

If the regular values have been lost, it restores them to default values. Then, if the TPM backup is valid, it updates the saved fields from the backup. That's the only time it uses the backup data.

The TPM backup is only written when specifically requested by crossystem, or when switching to normal mode. It reads the TPM first, so it doesn't write the TPM if the backup space already has the correct values.

The backup space is locked with PP, at the same time as the kernel rollback space.

Patch-set: 2