Update patch set 2
Patch Set 2:

Nope. This is a change entirely in this one repo. The rambi builder is trying to link host/linktest/main and host/linktest/extern BEFORE it's built the (local) libraries on which they depend, despite the dependencies being clearly listed in the Makefile:

uhh. 

Well, they *were* in the Makefile.

Poo. I accidentally took 'em out in an earlier commit. I told Don it was something stupid, and I'd have to make a public fool of myself in order to find it.

Stand by for another revision...

Patch-set: 2