Update patch set 1
Patch Set 1: Code-Review+1

After discussion with Nam, most of my complaints were due to these lines in cgptlib_internal.h:

/* Defines GPT sizes */
#define GPT_PMBR_SECTOR 1  /* size (in sectors) of PMBR */
#define GPT_HEADER_SECTOR 1

I assumed that "GPT_HEADER_SECTOR" was the address of the GPT Header. From the comment, it's supposed to the size of the GPT Header in blocks.


If that last macro was changed to be "GPT_HEADER_SECTORS" instead, then I agree with what Nam has written.

We should still add a "TODO" comment pointing out that the secondary header doesn't have a way to recreate the gap between the primary header and primary table

Unless we put that same gap in the secondary structs too. Then it would be able to calculate the required gap. That would be okay with me.

Patch-set: 1
Label: Code-Review=+1