Use "futility FOO" in the scripts/ directory too
We still create the symlinks (FOO -> futility), but this
change makes all the scripts under scripts/ invoke the
built-in functions with "futility FOO ..." instead of using
the FOO symlink.

BUG=chromium:231547
BRANCH=ToT
TEST=make runtests

In addition to running "make runtests", I temporarily
modified the Makefile to avoid creating the symlinks at all.
The tests still passed, but we don't have any tests (or even
instructions) for things in the scripts/ directory, so that
means little.

Change-Id: I31a1c27348be50a138776e6c2c7f1f33f9f231d0
Signed-off-by: Bill Richardson <wfrichar@chromium.org>