Add hwid digest field to GBB header
This adds a field in the GBB header for either a sha1 or sha256
digest of the HWID string, and a field to indicate which one is
used. Tests and changes to gbb_utility are included.

BUG=chromium:415227
BRANCH=ToT
TEST=make runtests

Since the GBB is in the RO firmware, there should be no side
effects for existing devices (but even without that, they should
handle a minor version change without complaint).

Change-Id: Icdb2a0b564677b0b65e58df897d2ec5af3964998
Signed-off-by: Bill Richardson <wfrichar@chromium.org>