crossystem: Normalize GPIO value read from sysfs
The recent kernels may return a positive value to indicate that a
GPIO is high, but that value may not always be exactly 1.

In particular with the recent intel chipset gpio driver it returns
the bit that is set in the GPIO register.

BUG=chrome-os-partner:32645
BRANCH=samus,auron
TEST=build crossystem for samus, check that wpsw_cur is correctly
reported when WP screw is installed.

Change-Id: I97dc3040f350e1bf7159fc3df00dc47252380427
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>